move functions to better locations
